[CMPLRLEGCY-3323] Optimization improvements for assignment reductions.
This change implements some extensions to Loop Carry Collapse to better
optimize certain classes of conditional assignment reductions. The main
one is for reductions that reach seqopt in this form:

  int x = 0;
  for (int i = 0; i < n; ++i) {
    x = merge(cond(i), x, f(i));
  }

This is similar to the existing "OpMerge" case, but here the op doesn't
depend on the carried value of x and may even be a literal. This is
handled by the new Section type "Merge".

However, LLVM will get fancier for some cases involving i1 values and
generate them as an and/or reduction instead:

  bool x = false;
  for (int i = 0; i < n; ++i) {
    x = x || cond(i);
  }

In this case, the or1 (/and1) can be thought of as also being equivalent
to a merge1 with a literal value (1 for or1, 0 for and1) and so can be
optimized similarly. This case is handled by the new Section type
"Logical".

While testing these changes, I discovered a latent bug in Loop Carry
Collapse where incorrect code is generated for multiple Regions
terminating in merge Sections in the same group because the merge-if
conversion is done while generating the predicate stream and so is only
done for the first Region. I've fixed this by converting the
transformInner*Merge* functions to fixupInner*Merge* functions that are
run for all regions in a group.

I've also added one small peephole to CSADataflowCanonicalization which
converts picks of these forms:

  pick x, cond, 0, 1
  pick y, cond, 1, 0

To equivalent movs/nots:

  mov1 x, cond
  not1 y, cond

This helps eliminate some more extraneous operations in some relevant i1
reductions.

Change-Id: I20f2e263141f17d3e27eb044769adcb60ccb3512
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/243759
Reviewed-by: Chen, Chih-Ping <chih-ping.chen@intel.com>
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>