[VPlan] Drop unknown metadata from vector memory references in HIR vectorizer.
Metadata kinds such as nonnull, dereferenceable are not accepted with a vector
of pointers. The issue causes compiler self builds to fail in xmain-web. The
changes here remove metadata kinds that we do not know how to propagate to
the generated vector memory references.

VPlan LLVM-IR vectorizer does not preserve metadata currently. When we start
preserving metadata in the same, we need to make sure that we delete metadata
such as nonnull.

Jira: CMPLRLLVM-19103.

Change-Id: Id7b9ae65b0327bf13d9cd6ff3b207531bd04f4cf
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/244211
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Reviewed-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Tested-by: Gaba, Abhinav <abhinav.gaba@intel.com>