[VPlan][NFC] Fix Klocwork reported issues.
This patch applies asserts for the multi-exit loop transformations in VPlan to
prevent nullptr dereferences in cases where there are potentially multiple loop
latches for inner loops. We have legality checks in place for the outer loop in
VPlan, but do we also need something similar for inner loops? We may have many
other assumptions about inner loops in VPlan.

Change-Id: Idf5603aa0eb53f19f3c7c4d7d01c2590d387dee5
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/244515
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>