[VPlan] Fix DA to propagate branch divergence correctly.
This patch fixes DA to properly propagate branch divergence when the
CondBit exists in a different block than its correspoinding instruction.
Propagation is now updated to start from the block containing the CondBit
since this is the actual source of potential divergence. DA previously
assumed that the instruction and CondBit were in the same block. DA now
keeps a map of CondBits and their corresponding blocks, so that when the
main compute() method visits instructions, the map is queried to see if
the instruction is a CondBit, and if so, begins propagation from the block
in the map.

Change-Id: Ie6b394b1680272193fe6f44af56744e311731b33
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/244544
Reviewed-by: Masten, Matt <matt.masten@intel.com>
Tested-by: Masten, Matt <matt.masten@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>