Update for the whole program analysis trace
This change set updates the trace of the whole program
analysis. It prints the result of each of the
requirements to achieve whole program:

  WHOLE PROGRAM RESULT:
    MAIN DEFINITION:  DETECTED
    LINKING AN EXECUTABLE:  NOT DETECTED
    WHOLE PROGRAM READ:  NOT DETECTED
    WHOLE PROGRAM SEEN:  DETECTED
    WHOLE PROGRAM SAFE:  NOT DETECTED

The result now prints if "main definition", "linking an
executable", "whole program read,seen" and "safe" were
"detected" or "not detected". It can also print if
whole program assume was enabled.

Change-Id: I4eeccfa99808f78c51a9aa33c0d4499a71143aba
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/244785
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>