[NFC][VPlan] Improve checks for external defs in VPlan check generation tool
Changes here implement the following in the tool used to update VPlan checks.

1) Since the print order of external defs is non-deterministic, use CHECK-DAG
to check for the same.
2) Avoid variable substitution in the value part of such external defs. This
avoids the problem of manually fixing up the order of generated check-dags.
Consider the checks:

; CHECK-DAG:   [[VP0:%.*]] = [[N30:%.*]]
; CHECK-DAG:   [[VP3:%.*]] = [[N10:%.*]] + [[N20:%.*]]

The order here has to be manually modified to check for N30 after N10 + N20.
Otherwise, the pattern for N30 will also match N10 + N20 pattern causing the
later check to fail. The checks now look like the following which addresses
such problems.

; CHECK-DAG:   [[VP0:%.*]] = %n3
; CHECK-DAG:   [[VP3:%.*]] = %n1 + %n2

Change-Id: I9cc38a93cb7420d20f54bbdf3b6e440fa5b1819e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/245569
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>