[VPlan] Avoid invariant decomposition by default during VPlan HIR construction
The changes in this patch set avoid decomposing invariant canon expressions
and blobs during VPlan construction of HIR loops. This change is now on by
default. Also included are changes to do regular decomposition for single
IV only canon expressions and canon expressions containing a single blob.
Several LIT tests have been modified to reflect changes in decomposition.
When printing the values of external defs, the values printed are embedded
in {} to avoid having to manually update the generated checks for external
defs. This avoids issues such as the following, where we need to place the
check for %m + -1 before the check for %m for the test to be reliable.

; CHECK-DAG:     [[VP5:%.*]] = %m
; CHECK-DAG:     [[VP0:%.*]] = %m + -1

Change-Id: I57bcf123af9de604ed4d98916871bc089c0d8db0
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/245683
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>