Improve entry guard/ztt recognition in ScalarEvolution by cancelling out positive constants in predicates. Addition of 1 to both sides commonly happens in ScalarEvolution during analysis so cancelling them out helps.
For example, for the following loop-

i =t;
while (i > 0)
 i--;

We need to prove (1 + zext(t))<nuw><nsw> >s 1

Using found predicate-
sext(t) >s 0

This becomes easier when we cancel 1 from both sides of the predicate to be proven.
The constant difference logic doesn't work because of mismatch in zext()/sext().

Change-Id: Id9e18e65881ff7c84b66b208394097a9e6001bea
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/245847
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>
Tested-by: Chawla, Pankaj <pankaj.chawla@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>