[DPCPP][HOST] Port of DataPerBarrier pass over from OpenCL
This is a part of effort to enable DPCPP kernels on the HOST through
WGLoopCreator/Barrier approach.

The goal is to enable a subset of benchmarks by doing this, not provide
product quality from day #1.
This port does not aim to port complete functionality from OpenCL to
DPCPP, only core functionality that is needed achieve the above.
So unless problems detected in review are critical to this goal, I'd
rather file JIRAs and address them later.

The pass is a prerequisite of the main Barrier pass.

Change-Id: I3060da2c1f2638da8b0db6067f6423eb589ee36f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/245943
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>