Update Transforms/LoopVectorize/novect-lcssa-cfg-invalidation.ll LIT test to disable unexpected optimizations.
The test was added in:
80cd36e [Test] Add a test showing how CFG analyses are invalidated after LV
And later updated in:
9cd4deb [LoopVectorize] Preserve CFG analyses if CFG wasn't modified
(https://reviews.llvm.org/D78360)

The purpose of the test is to check whether or not CFG was preserved after loop-vectorize.

It adds jump-threading after loop-vectorize in the opt command expecting to
be a no-op, but JumpThreadingPass makes optimizations in xmain that makes
the final IR different.

This change is to disable those optimizations through the opt command in the test.

Change-Id: I229d7a116c824b77bea421970e29800f46e25ec8
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/246046
Tested-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>
Reviewed-by: Gaba, Abhinav <abhinav.gaba@intel.com>