[VPlan][2/4] Explicitly represent and propagate operator attributes in VPlan.
This patch introduces a new mechanism in VPlan framework to explicitly
capture and represent operator-related attributes like FastMathFlags,
OverflowingFlags and ExactFlag. This is achieved by using tagged unions
to implement differentiation among operator types. The patch also
updates VPlan LLVM-IR vectorizer to correctly import and emit these
attributes in outgoing vectorized LLVM-IR. This is the second patch in
the series for the task to represent underlying attributes/metadata
information in VPlan.

JIRA : CMPLRLLVM-11656

Change-Id: Ib69f202051b46ce9d87ce63237acec2c103e735a
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/246263
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Tested-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>