[VPlan] Remove assert in VPOCodegen::serializeInstruction().
This assert seems to be a remnant from the older codegen and seems
unnecessary. The file
https://git-amr-2.devtools.intel.com/gerrit/#/c/164542/9/lib/Transforms/Vectorize/Intel_VPlan/IntelLoopVectorizationCodeGen.cpp
has the assert and is meant to check for "vector type" and doesn't seem correct.

This change itself does not cause the failure when SPEC2000, SPEC2006C,
SPEC2000Cpp are run in stress mode.

Change-Id: Ifdea1412d98cb7754320ca4ced9497c883b111fe
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/246359
Reviewed-by: Pushkar Ratnalikar <pushkar.v.ratnalikar@intel.com>
Tested-by: Pushkar Ratnalikar <pushkar.v.ratnalikar@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>