[VPO][OpenMP] Fix the issue for reduction with __m128 type
__m128 is emitted to <4 x float> vector type in IR. Pass this type to
genReductionScalarInit which returns the value with vector type to
initialize the local reduction variable. The exception is raised if
scalar type (float) is passed, since the value returned from
genReductionScalarInit, is with scalar type and its type is not same
as the vector type for local reduction variable.

JIRA: CMPLRLLVM-8445

Change-Id: Ifa494bffb84e4e4e1be6667b470358ea48977da8
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/247007
Reviewed-by: Zou, Feng <feng.zou@intel.com>
Tested-by: Zou, Feng <feng.zou@intel.com>