[VPlan][HIR] Use VPHIRCopyInst to represent copy instructions in VPlan.
HIR framework was recently updated to use llvm.ssa_copy intrinsic as
underlying instruction to represent copies in HIR. When we decompose
HIR to VPlan IR, we want to use the specialized VPHIRCopyInst to
represent copies instead of VPCallInst (due to underlying intrinsic).
This patch updates VPlanHIRDecomposer and VPlan framework to implement
this change. Additionally this patch also fixes decomposer to not
emit unnecessary VPCallInst for standalone addressof refs coming from
Fortran-based HIR.

Change-Id: Ief997585fe8ca5aab9eaab43e00b361e86848442
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/246933
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Tested-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>