CMPLRLLVM-19454: Prevent dtrans::ResolveTypes for types used in dllexport function
The following change set prevents that dtrans::ResolveTypes merge types
that are used in a dllexport function. The issue is that there is a chance
that these functions will be used by an external library. If these functions
are removed then it will produce a missing symbols error during link time.

Also, the types that are used as parameters or return types in a dllexport
function will be marked with the safety issue System object.

Change-Id: Ia8a5fa1efb6f3e4470e85958416c8970fe9b4cbd
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/247191
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Tested-by: Rivera, Axel Y <axel.y.rivera@intel.com>