[VPlan] Preserve uniform control flow by default in HIR vectorization.
Designated reviewers -- Primary: Matt, Secondary: Karthik

The changes here are needed to address CMPLRLLVM-10389 which is on beta07 MFL.
The changes were disabled earlier due to a stability issue in QA testing which
is now resolved.

Search loop vectorization has not been uplifted fully to explicit representation
and the idiom recognition is picky. The change is thus disabled for search
loops.

Jira: CMPLRLLVM-10389

Change-Id: I0f273c6cfa06b574a8ea157c680107c524164050
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/247384
Reviewed-by: Masten, Matt <matt.masten@intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Tested-by: Guggilla, Satish <satish.guggilla@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>