[DPCPP][HOST] Port of DataPerValue pass over from OpenCL
This is a part of effort to enable DPCPP kernels on the HOST through
WGLoopCreator/Barrier approach.

The goal is to enable a subset of benchmarks by doing this, not provide
product quality from day #1.
This port does not aim to port complete functionality from OpenCL to
DPCPP, only core functionality that is needed achieve the above.
So unless problems detected in review are critical to this goal, I'd
rather file JIRAs and address them later.

The pass is a prerequisite of main Barrier pass.

Change-Id: If2e235069b69d835dafd72ea56b494bd1292354c
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/247313
Reviewed-by: Wenju He <wenju.he@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>