[VPlan] Use CanonExpr inteface for standalone IV check in HIR decomposition.
Designated reviewers -- Primary: Karthik, Secondary: Sergey K

When checking a canon expression for standalone iv such as [i1], we missed
checking for constant additive and denominator. Using canon expression
interface fixes the same. Also, when trying to build invariant external defs
after removing the IV, we force denominator to 1 and set destination type
to be the same as source type. When doing the check to see if we should do
regular decomposition with IV removal, we now modify the CE in the same way
to keep both in sync.

Change-Id: I70dcd5a1c10bfb99c78ab9e1a71d6eb24a46a237
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/247375
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Tested-by: Guggilla, Satish <satish.guggilla@intel.com>