CMPLRLLVM-19321: Fix Sema crash when rejecting unknown TS builtin
About 5 years ago, community added support for an aux-triple.  In doing
so, they added the Target Specific 'builtin' list from the aux-triple.
However, SemaChecking's target specific checks don't actually check the
aux-triple versions.  Typically, this isn't a problem as most
aux-triples don't actually have target specific builtins, or are used
with a primary triple that doesn't have target specific builtins.

In cases where the aux-triple and triple both have target specific
builtins, and the aux-triple is used, most targets don't actually reject
the builtin. A few actually fall out of their function (which is UB).
Therefore this hasn't been noticed.

However, the IntelFPGA code added a couple of builtins to the SPIR
target. Combined with the x86_64 aux-triple, we end up in the 'bad' case
where BOTH the aux-triple and triple have a builtin, and the aux-triple
version is called. The IntelFPGA version actively rejected unknown
builtins, so the return of ExprError caused the parameters to not have
their cleanups run.  The test itself shows a situation where a temporary
is created, but never cleaned up, leading to a frontend assertion.

This patch puts a temporary fix for this in that makes the IntelFPGA
builtin handling code silently accept unknown builtins (and refactors
the code slightly).

Change-Id: If14caf5eca4a71f4ade5a1f50551cc3e0d084f28
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/247525
Reviewed-by: Keane, Erich <erich.keane@intel.com>
Tested-by: Keane, Erich <erich.keane@intel.com>