CMPLRLLVM-19245: Revised qsort recognition using Function Recognition function pass.
This should be the final initial development change set for qsort recognition.
Functions that are matched by the Function Recognizer get an attribute with
the name starting with is-qsort. When one of those Functions calls another
Function, in the case of a direct call, the Callee is marked with an attribute
whose name starts with 'must-be'. In the case of an indirect call, the call
itself is marked with the 'must-be' attribute, as we don't know what will be
called until indirect call conversion is performed.

The inliner is augmented with a function in Attributes.cpp which checks that
each Callee of the Caller which is being inlined has matching 'must-be*' and
'is-qsort*' attributes.  If it doesn't, function recognition is disabled for
the caller.

At qsort recognition time, calls and callees are again matched up to see
whether the Function Recognition can be preserved. The code is much smaller
than it was originally, as most of the recognition work is now done in the
Function Recognizer. The tests in the Intel_QsortRecognizer directory have
been replaced by just a few tests.  Don't be confused by Gerrit's attempt to
match up some of the deleted tests with new tests that I added.

Change-Id: I807578a198aaf41ff3ad3aaa4974c35b9b61d6c0
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/248344
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>