[NFC][VPlan][VPBasicBlock] Keep track of block-predicate instead of its operand
This makes VPValue::replaceAllUsesWith just work, without any special handling
for the block-predicate uses where the VPBB has to be updated as well.

This patch doesn't fix a similar issue for splice and block-predicate's parent
but puts asserts in place to guard against that one as well. That second issue
is much rarer though.

Change-Id: Ic66767de55b669a3d2e5918fd820aa37ea46d30d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/248110
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: Elovikov, Andrei <andrei.elovikov@intel.com>