CMPLRLLVM-19657: Fix for descaling sums involving casts
Descale will descale sums if both addends are tested to be descalable
using the "TestOnly" parameter. After testing but before attempting to
descale the addend it creates the new descaled sum instruction as an
additional user of the addends so that any other users of the descaled
operands are not incorrectly affected.

Descale has special logic to descale BinaryOps with multiple users, but
not Casts. The consequence of this is (as seen in CMPLRLLVM-19657) is
that Descale will claim that it can descale the cast, but then fails due
to the second use by the new descaled sum operation.

To resolve this, this change extends the handling for BinaryOps with
with multiple users to Cast instructions with multiple users.

Change-Id: Iadcc37168b0aa4315fe73183ad412153339a5fc6
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/248846
Reviewed-by: Bakhvalov, Denis <denis.bakhvalov@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>