CMPLRLLVM-5594: Extend local pointers analysis for zero element access
This change set introduces an analysis in the local pointers
analysis to catch a zero element access for nested structures.
For example:

  %"class.outer" = type { %"class.inner" }
  %"class.inner" = type { %class.TestClass*, %class.TestClass*}
  %class.TestClass = type { i64, i64, i64}

Consider the following BitCast and Load instructions, where %0
is a GEP or an Argument:

   %1 = bitcast %"class.outer"* %0 to i64*
   %2 = load i64, i64* %1

Instructions %1 and %2 basically means that is loading the element
0 in %"class.inner". If that is the case then we add %class.TestClass
as part of the aliases to %2. Also, this pattern will be treated as a
read of element 0 in %"class.inner".

Change-Id: I6c9da4c5b0e9b29fe3c9e2c165014f1c57f93e75
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/249069
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Tested-by: Rivera, Axel Y <axel.y.rivera@intel.com>