CMPLRLLVM-19602:Allow extra void* param in target variant dispatch variants
When a 'target variant dispatch' variant is seen with one extra
void* argument, check the compatiblity without it.  If they are
compatible allow it.  Otherwise do the normal checking.

Note that the normal C++ case tries an implicit conversion and might
match a slightly different set.  In the case with an extra parameter
this would always fail so a simpler mergeFunctionTypes() is used
instead.

Change-Id: Icc85a66bea42bcf95654f115a659ad407f3202b5
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/248887
Reviewed-by: Rice, Michael P <michael.p.rice@intel.com>
Tested-by: Rice, Michael P <michael.p.rice@intel.com>