[VPlan] Bailout of VPlan for unsupported AggregateType instructions.
Instructions producing AggregateType like structs and array cannot
be vectorized directly, they should be emulated via serialization.
Some components of VPlan like CostModel and Codegen need to be
aware of this and fixed appropriately to avoid stability issues.
This patch adds a temporary bailout mechanism as a placeholder
until this issue is resolved throughout VPlan. Additionally we
need to erase the AUTO.VEC directives surrounding the loop if
VPlan decides to bailout early from vectorization.

Change-Id: Ic36d3705157ab3cddb534feaa4969afd1a193f0a
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/248879
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>