[CMPLRLLVM-19888] Add a DAG combine peephole for the satd_8x4 routine 525.x264.
The satd_8x4 function uses the upper and lower 16 bits of 32 bit elements
as almost separate values using adds and subtracts. But they
allow carries/borrows from the low 16-bits into the high 16 bits.

This is eventually cancelled out with an absolute value if the
lower 16 bits represent a negative number. The code ends up adjusting
the upper bits to get rid of the extra bits.

Due to this we can't use 16-bit vector elements, but we can still
play some tricks using byte and word shuffles to setup the data.
Then we can bit cast to the vXi32 vector once the 16 bit values are
in place.

Unfortunately, its a pretty large pattern match to get from the
add that puts the low and high 16 bits together all the way back
to the four v8i8 loads that supply the data.a

Instead of splitting the four loads and rejoining them into 4
vectors, doing 4 extends, and 2 separate subtracts before
combining with shift and add. We combine the loads using unpck
and pshufb. Do 2 extends to vXi16 and then do a single subtract.
Reducing the attached test by 5 instructions. The whole pattern
occurs twice in the benchmark, but I've omitted that for the test.

Change-Id: If481616796c534130f8de4a21daae1d0698c952b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/248389
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>