[VPlan] Miscellaneous GEP fixes in VPValue code generation for HIR
Designated reviewers -- Primary: Karthik, Secondary: Matt

Changes here fix the following issue related to generating code
for GEPs and loads/stores in VPValue based code generation for
HIR.

1) If a GEP instruction has a single operand, use the ref corresponding
to the operand as the ref for the GEP instruction as well.

2) The base canon expression of an addressof ref needs to be a self-blob.
If the base ref is not a self blob, create a copy instruction and use
the lval of the copy instruction to form the new addressof ref.

3) When forming a memref such as blob[0] and we are dealing with a unit
strided ref, the type of 0 needs to remain scalar.

Change-Id: I622dd50fd0b8b1ccb4fda04e1bb7ab260ee388f1
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/249169
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>