[NFC][VPlan][HIR] Add couple of debug cl::opts
Since VPlanDriver[HIR] is a function-level pass it is hard to map some VPlan
dump/debug output to a particular loop, especially in the HIR case when we don't
have any names for VPValues. To ease that mapping during debuggin add options to
dump the incoming HLLoop for each HLLoop processed and the same for the
vectorized HIR output. Note that we already have the latter under
-debug=<option> but it enables much more than just that.

Change-Id: I8cbc78d6c45bd89eed894f05e8748355fd8b44f2
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/249102
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: Elovikov, Andrei <andrei.elovikov@intel.com>