[SYCL] Fix host-to-host copy during copyback (#1692)
Copyback is performed unconditionally during buffer destruction, so
it's possible for it to perform a copy from and to the same host
pointer. This patch adds a check for this scenario so that we don't rely
on undefined behaviour of std::memcpy.

Signed-off-by: Sergey Semenov <sergey.semenov@intel.com>

This is a cherry-pick of intel/llvm PR #1692:
https://github.com/intel/llvm/pull/1692

It fixes the following issue:
https://jira.devtools.intel.com/browse/CMPLRLLVM-15468

Change-Id: I4aa7b1e2de3b100c86321600f325f244637ae252
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/249313
Reviewed-by: Sergey Kanaev <sergey.kanaev@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>