[VPlan][HIR] Unmask operations other than load/store/call in VPValue based code generation
Designated reviewers -- Primary: Karthik, Secondary: Matt

Safe operations such as multiply need to remain unmasked in VPValue code
generation. This is important for stability as well as performance of
VPValue CG. In mixed CG mode, we can have multiple HLInsts that write to the
same temp under different mask. Temps in HIR correspond to memory locations
and an HLInst such as 't1 = val @Mask = m1' is lowered as:

  ldval = load from t1 memory location
  blendval = select m1, val, ldval
  store blendval, t1 memory location

This sequence is unnecessary in VPValue CG where different VPInstructions
write to different temps(except in known cases such as copy instructions
resulting from SSA deconstruction or in case of reductions). This also can
actually cause stability issues. Consider the following scalar HIR:

   DO i1 = 0, 99, 1   <DO_LOOP> <simd>
     %1 = (%unif)[0];
     %2 = (@larr)[0][i1];
     if (%2 != 0)
     {
         (@larr2)[0][(%n2 * %1)][i1] = i1;
     }
   END LOOP

The value %n2 * %1 is uniform and the store to larr2 is unit strided. When
trying to form the unit strided ref, we use scalar value for lane 0 for this
uniform value. If the multiply remains masked, the value in this lane may
be garbage if the mask for lane 0 is false. For operations other than calls,
loads/stores, we need to generate unmasked operations for VPValue CG. This
is essentially the same as what we do for LLVM IR path.

Change-Id: I1700ba862807ccf9631e6021c62e5c766a45a96f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/249293
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Tested-by: Guggilla, Satish <satish.guggilla@intel.com>