[VPlan][Predicator] Do not transform phi to blend after uniform branch
The algorithm for calculating live-in blocks for merge-phi placement in
phi-to-blend processing relies on being part of the linearized chain. If that
assumption doesn't hold we might go over the backedge adding unnecessary live-in
blocks for which no blending values exist, failing like

  Assertion `BBToBlend.size() > 0 && "No values for blend!"' failed.

The whole special casing/workarounding should be dropped once CG stops blocks
merging and won't require ask to create unnecessary blends in the first place.

Change-Id: I403e786a186b68fd37355e1a494b9ad5bf795333
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/249398
Reviewed-by: Masten, Matt <matt.masten@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>