[Driver] Improvements to -flto behaviors
When dealing with -flto, we want to pass along any -mllvm options we are
using with the regular compilation to the link step as well.  Created a
more centralized function that can be used to add these.

Also hooked up setting the optimization level for lld-link on Windows.

When using -Qipo, enable -fuse-ld=lld automatically

Change-Id: Ie0b0d4d5cfffe13d5e5cc41a1626b1d2a3116d92
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/249742
Reviewed-by: Toguchi, Michael D <michael.d.toguchi@intel.com>
Tested-by: Toguchi, Michael D <michael.d.toguchi@intel.com>