[HIR][VPlan] Recognize and handle SIMD loops with directives in preheader/postexit.
SIMD directives in a HIR region can be embedded in a HLLoop's
preheader/postexit blocks. The current assumption in HIR framework
and VPlan is that they are sibling nodes to the loop. This patch
updates HLLoop's SIMD region detection and VPlan to handle this
scenario and vectorize such loops via SIMD vectorization instead
of auto-vectorization. We also take this opportunity to cleanup
and older implementation of VPlanDriverHIR::eraseLoopIntrins().

JIRA: CMPLRLLVM-19929

Change-Id: I037e6029d4ca2bae0a625ed91ba64e38f1a3a6e2
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/249762
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>