[VPlan][Loop Exits Canonicalization] Stop forcing full linearization
Strictly speaking, this is a functional change that removes the full
linearization that was committed as a workaround back in time when the loop
exits canonicalization didn't preserve SSA form. Later, that was fixed and the
fix came with its own LIT tests verifying that SSA form is preserved as needed.

However, that fix still kept the workaround of forcing full linearization in
place, even if it isn't needed anymore. I don't think such workaround removal
needs a dedicated LIT test because it won't make any sense without the legacy
context in it - it won't be covering *any* existing code. Instead, this is
supposed to be tested via end-to-end tests - performance benchmarks.

Change-Id: Ic9f175367e9e568b090b25adaa65838fc8752816
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/249761
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Mitropoulou, Konstantina <konstantina.mitropoulou@intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>