[DTrans - OpaquePtr] Add pointer type collection for element-zero loads
A pointer to a structure can be used to directly load the first element
of the structure without first getting the address of field using a
GEP instruction. This updates the way load instructions are processed
where an element is loaded from a pointer to a structure to collect
this as a load of the structure element contained there (including
consideration for nested types), rather than storing the structure
type itself in the pointer alias set.

Change-Id: I0a042f90c93c4487332a770eb65de685630e137b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/249947
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>