Fix aostosoa-transform-safety01.ll fail on xmain-web due to community commit “ [NewPassManager] Add assertions when getting statefull cached analysis.”
New PM is not invoking DTransAnalysis even if a pass doesn’t preserve the DTransAnalysis. I don’t find any better way to force new PM to invoke DTransAnalysis again when a pass make changes to types. For now, I add it to only AOSToSOA.cpp to make the LIT test pass but need to be added to other passes on xmain later.

Change-Id: I3abd981c8f4bc8dce6cf7c76aed62209be037cdd
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/250031
Reviewed-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Tested-by: Gaba, Abhinav <abhinav.gaba@intel.com>