CMPLRLLVM-20050 - Add default constructor for PragmaLoopHintInfo.
Users of PragmaLoopHintInfo always need to initialize the two Token
fields (PragmaName and Option) since Token is not default constructed.
That is usually done by copying a local Token, but since Option is not
always used it is easy to forget to initialize it with startToken().

The default constructor lets us only deal with Option when needed in
the particular pragma.  Unlike the fairly simple uses in community
pragmas, some of ours are fairly complicated making it more likely
to forget handling Option manually (causing sanitizer errors).

Also added a contructor that takes the PragmaName that initializes
both Tokens and changed most of our pragmas to use it.

Change-Id: Id19a831e2c3ee707cdc023474183ca34aad8c12c
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/250524
Reviewed-by: Keane, Erich <erich.keane@intel.com>
Reviewed-by: Rao, Premanand M <premanand.m.rao@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Yu, Jennifer <jennifer.yu@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>