CMPLRLLVM-18458 - Fix EmitFakeLoadForRetPtr so ReturnExpr is emitted once
Previously the code tried to load from the lvalue and check for a load
instruction.  If there was no load instruction it fell back to the original
code and emitted the return value again.

Now EmitFakeLoadForRetPtr will always return the correct value when the
fakeload optimization case is possible, even when it doesn't add a fakeload.

Change-Id: I1926fae24bb230104d15c5881e967c57975bdc39
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/250625
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Keane, Erich <erich.keane@intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>