[SYCL] Always build gtest for SYCL with stdc++
The problem is that when LLVM_ENABLE_LIBCXX=ON is enabled,
SYCL unittests and libsycl.so library have different mangling
for some functions. This happens because google_tests are compiled with
LLVM's c++ headers while libsycl.so compiled with GNU c++ headers.

This patch resolves the issue above by building gtest and its
dependencies twice and link each instance against proper standard library.

Change-Id: Ia8027c2dceedf198c0c5ed39fb5eae771c486c60
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/251239
Reviewed-by: Romanov, Vlad <vlad.romanov@intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>
Tested-by: Dai, Chunyang <chunyang.dai@intel.com>