[VPlan] Default initialize VPOperatorIRFlags based on operator kind.
MemorySanitizer caught that the VPOperatorIRFlags object in every
VPInstruction is uninitialized and can contain junk when new
VPInstructions are created on-the-fly during VPlan-to-VPlan
transforms. The flags are tested and copied over in vectorizer
codegen based on instruction's operator kind. This patch adds
default initialization of VPOperatorIRFlags based on the operator
kind that instruction corresponds to.

JIRA : CMPLRLLVM-20053

Change-Id: I2026be60a97d2fa6b3b58723f790b244ac4256ec
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/251242
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>