[X86] Enable the global FMA pass for 128/256-bit vectors when only AVX512F is enabled.
Previously we would only look for EVEX or VEX operations
depending on whether avx512f was enabled, but 128/256 would
still be using VEX opcodes avx512vl was not also enabled. Fix
this by looking for VEX and EVEX opcodes. But when generating
FMA instructions take vector width, feature flags, and vector
width into account.

Also fix an incorrect subregister extract in the broadcast
handling for materializing a 1.0 constant. We were using
a method that is for getting the subreg of a phyiscal register,
but we're working on virtual registers so we need to generate
a copy instruction.

Change-Id: Ieabe3f8dd06058deab1fc5dc81719221e3775078
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/251355
Reviewed-by: Topper, Craig <craig.topper@intel.com>
Tested-by: Topper, Craig <craig.topper@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>