[DTRANS] CMPLRLLVM-19976: Prevent visiting GEP instructions in dead blocks
This change set adds the dominator tree analysis to DTrans for checking
if a basic block is dead. If the dominator tree analysis finds that the
basic block is dead, then it won't visit the GEP instructions that are
in it. This changes are for preventing a possible infinite loop in DTrans
analysis when a GEP is in a broken SSA form, but is inside a dead block.

Change-Id: I42ca241acab80a8505b0bb3c7e2b632c1a8457fb
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/252275
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>